YD-477 Refactored public key crypto tests
